requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
google-generativeai==0.3.2
Pillow==10.0.1
SQLAlchemy==2.0.43
gunicorn==21.2.0
